app_stack/app_backend/requirements.txt
fastapi[standard]
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
pydantic-settings